pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "vitac"
version = "0.1.0"
description = "Voice-Interactive Terminal Agent Collaboration Benchmark"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "typer>=0.12",
    "rich>=13.0",
    "docker>=7.0",
]

[project.optional-dependencies]
tts = ["gtts>=2.5"]
stt = ["openai-whisper>=20230918"]

[project.scripts]
vitac = "vitac.cli:app"
vb = "vitac.cli:app"

[tool.setuptools.packages.find]
include = ["vitac*"]
